delint, and 64bit fixes
git-svn-id: https://svn.aros.org/svn/aros/trunk@50011 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/libs/xad/amiga/source/test/CreateFilenameMask.c
@@ -57,6 +57,6 @@ void main(void)
   {
     if(!(i&7))
       printf("\n");
-    printf("0x%02lX%s", mask[i], i < 256/8-1 ? "," : "\n};\n");
+    printf("0x%02lX%s", (long unsigned int)mask[i], i < 256/8-1 ? "," : "\n};\n");
   }
 }
